requirements.txt
# route_finder.py
numpy ~= 1.25.0
scikit-image ~= 0.21.0
scipy ~= 1.11.0
opencv-python ~= 4.7.0.72
shapely ~= 2.0.2
networkx ~= 3.1

# app.py
Quart ~= 0.19.4
python-socketio ~= 5.10.0
Hypercorn ~= 0.15.0